requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
Babel==2.9.1
certifi==2023.5.7
charset-normalizer==3.1.0
frozenlist==1.3.3
greenlet==2.0.2
idna==3.4
magic-filter==1.0.9
multidict==6.0.4
openai==0.27.7
psycopg2-binary==2.9.6
pyTelegramBotAPI==4.12.0
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
SQLAlchemy==2.0.15
SQLAlchemy-Utils==0.41.1
tqdm==4.65.0
typing_extensions==4.6.3
urllib3==2.0.2
yarl==1.9.2